Making the code more readable by adding reusable components
