+++ apps/client/src/components/categoryItem/SeedData.ts
@@ -1,0 +1,7 @@
+
+export const categoryItemsSeed: CategoryItem[] = [
+    { id: 1, name: 'Laptop', description: 'High-performance laptop', price: 1000, stock: 10, pictures: Array.from({ length: 4 }, (_, i) => `https://picsum.photos/200/200?random=${i}`) },
+    { id: 2, name: 'Headphones', description: 'Wireless headphones', price: 200, stock: 20, pictures: Array.from({ length: 4 }, (_, i) => `https://picsum.photos/200/200?random=${i}`) },
+    { id: 3, name: 'Mouse', description: 'Gaming mouse', price: 50, stock: 30, pictures: Array.from({ length: 4 }, (_, i) => `https://picsum.photos/200/200?random=${i}`) },
+  ];
+
