+++ apps/client/src/components/ordersTable/OrdersTable.tsx
@@ -1,4 +1,6 @@
 import React, { useState } from 'react';
+import ReusableInput from '../ReusableInput';
+import OrdersTableSeed from './OrdersTableSeed';
 
 interface Order {
   id: number;
@@ -9,11 +11,7 @@ interface Order {
 }
 
 const OrdersTable: React.FC = () => {
-  const [orders, setOrders] = useState<Order[]>([
-    { id: 1, customerName: 'John Doe', email: 'john@example.com', totalPrice: 100, status: 'Pending' },
-    { id: 2, customerName: 'Jane Smith', email: 'jane@example.com', totalPrice: 200, status: 'Shipped' },
-    { id: 3, customerName: 'Alice Johnson', email: 'alice@example.com', totalPrice: 150, status: 'Pending' },
-  ]);
+  const [orders, setOrders] = useState<Order[]>(OrdersTableSeed);
   const [editOrder, setEditOrder] = useState<Order | null>(null);
   const [newOrder, setNewOrder] = useState<Order>({ id: 0, customerName: '', email: '', totalPrice: 0, status: '' });
   const [showAddOrderForm, setShowAddOrderForm] = useState(false);
@@ -72,68 +70,42 @@ const OrdersTable: React.FC = () => {
     <div className="orders-table">
       <h2 className="text-2xl font-semibold mb-4 uppercase text-gray-800">Orders Table</h2>
       <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-  <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
-    <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-      <tr>
-        <th className="px-6 py-3">ID</th>
-        <th className="px-6 py-3">Customer Name</th>
-        <th className="px-6 py-3">Email</th>
-        <th className="px-6 py-3">Total Price</th>
-        <th className="px-6 py-3">
-          <span className="sr-only">Actions</span>
-        </th>
-      </tr>
-    </thead>
-    <tbody>
-      {orders.map((order) => (
-        <tr key={order.id} className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
-          <td className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">{order.id}</td>
-          <td className="px-6 py-4">{order.customerName}</td>
-          <td className="px-6 py-4">{order.email}</td>
-          <td className="px-6 py-4">{order.totalPrice}</td>
-          <td className="px-6 py-4 text-right">
-            <button className="font-medium text-blue-600 dark:text-blue-500 hover:underline" onClick={() => handleEditOrder(order)}>Edit</button>
-            <button className="font-medium text-red-600 dark:text-red-500 hover:underline ml-2" onClick={() => handleDeleteOrder(order.id)}>Delete</button>
-          </td>
-        </tr>
-      ))}
-    </tbody>
-  </table>
-</div>
-
+        <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
+          <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
+            <tr>
+              <th className="px-6 py-3">ID</th>
+              <th className="px-6 py-3">Customer Name</th>
+              <th className="px-6 py-3">Email</th>
+              <th className="px-6 py-3">Total Price</th>
+              <th className="px-6 py-3">Status</th>
+              <th className="px-6 py-3">
+                <span className="sr-only">Actions</span>
+              </th>
+            </tr>
+          </thead>
+          <tbody>
+            {orders.map((order) => (
+              <tr key={order.id} className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
+                <td className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">{order.id}</td>
+                <td className="px-6 py-4">{order.customerName}</td>
+                <td className="px-6 py-4">{order.email}</td>
+                <td className="px-6 py-4">{order.totalPrice}</td>
+                <td className="px-6 py-4">{order.status}</td>
+                <td className="px-6 py-4 text-right">
+                  <button className="font-medium text-blue-600 dark:text-blue-500 hover:underline" onClick={() => handleEditOrder(order)}>Edit</button>
+                  <button className="font-medium text-red-600 dark:text-red-500 hover:underline ml-2" onClick={() => handleDeleteOrder(order.id)}>Delete</button>
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      </div>
       {editOrder && (
         <div className="mt-4 p-4 border border-gray-300 bg-white dark:bg-gray-800">
           <h3 className="text-lg font-semibold mb-2 text-gray-900 dark:text-white">Edit Order</h3>
-          <label className="block mb-2">
-            Customer Name:
-            <input
-              type="text"
-              name="customerName"
-              value={editOrder.customerName}
-              onChange={handleInputChange}
-              className="form-input mt-1 block w-full"
-            />
-          </label>
-          <label className="block mb-2">
-            Email:
-            <input
-              type="email"
-              name="email"
-              value={editOrder.email}
-              onChange={handleInputChange}
-              className="form-input mt-1 block w-full"
-            />
-          </label>
-          <label className="block mb-2">
-            Total Price:
-            <input
-              type="number"
-              name="totalPrice"
-              value={editOrder.totalPrice}
-              onChange={handleInputChange}
-              className="form-input mt-1 block w-full"
-            />
-          </label>
+          <ReusableInput label="Customer Name" name="customerName" value={editOrder.customerName} onChange={handleInputChange} />
+          <ReusableInput label="Email" name="email" value={editOrder.email} onChange={handleInputChange} />
+          <ReusableInput label="Total Price" type="number" name="totalPrice" value={editOrder.totalPrice} onChange={handleInputChange} />
           <label className="block mb-2">
             Status:
             <select
@@ -166,36 +138,9 @@ const OrdersTable: React.FC = () => {
       {showAddOrderForm && (
         <div className="mt-4 p-4 border border-gray-300 bg-white dark:bg-gray-800">
           <h3 className="text-lg font-semibold mb-2 text-gray-900 dark:text-white">Add Order</h3>
-          <label className="block mb-2">
-            Customer Name:
-            <input
-              type="text"
-              name="customerName"
-              value={newOrder.customerName}
-              onChange={handleInputChange}
-              className="form-input mt-1 block w-full"
-            />
-          </label>
-          <label className="block mb-2">
-            Email:
-            <input
-              type="email"
-              name="email"
-              value={newOrder.email}
-              onChange={handleInputChange}
-              className="form-input mt-1 block w-full"
-            />
-          </label>
-          <label className="block mb-2">
-            Total Price:
-            <input
-              type="number"
-              name="totalPrice"
-              value={newOrder.totalPrice}
-              onChange={handleInputChange}
-              className="form-input mt-1 block w-full"
-            />
-          </label>
+          <ReusableInput label="Customer Name" name="customerName" value={newOrder.customerName} onChange={handleInputChange} />
+          <ReusableInput label="Email" name="email" value={newOrder.email} onChange={handleInputChange} />
+          <ReusableInput label="Total Price" type="number" name="totalPrice" value={newOrder.totalPrice} onChange={handleInputChange} />
           <label className="block mb-2">
             Status:
             <select
@@ -238,3 +183,5 @@ const OrdersTable: React.FC = () => {
 };
 
 export default OrdersTable;
+
+
