+++ apps/client/src/components/ordersTable/OrdersTableSeed.ts
@@ -1,0 +1,5 @@
+export const OrdersTableSeed: OrderSeed[] = [
+    { id: 1, customerName: 'John Doe', email: 'john@example.com', totalPrice: 100, status: 'Pending' },
+    { id: 2, customerName: 'Jane Smith', email: 'jane@example.com', totalPrice: 200, status: 'Shipped' },
+    { id: 3, customerName: 'Alice Johnson', email: 'alice@example.com', totalPrice: 150, status: 'Pending' },
+  ];
