+++ apps/client/src/components/usersTable/userSeeds.ts
@@ -1,0 +1,6 @@
+export const usersSeed = [
+    { id: 1, name: 'John Doe', email: 'john@example.com', role: 'Admin' },
+    { id: 2, name: 'Jane Smith', email: 'jane@example.com', role: 'User' },
+    { id: 3, name: 'Alice Johnson', email: 'alice@example.com', role: 'User' },
+  ];
+
